ucorrelate with floats
Changing `zeros` to `empty` only speeds up the function a little, as it avoids filling C with zeros before filling it with the correct values, and could also be applied to lines 79, 82 and 84.
Changing `np.int64` to `t.dtype` makes the output the same type as the input. It can still go wrong it `t` has ints and `u` floats, but taking that into account can be left to the user.
Added the example to the documentation.